added all states and territories

+++ src/main/java/io/catalyte/training/sportsproducts/domains/purchase/ShippingRate.java
@@ -14,6 +14,6 @@ public class ShippingRate {
 
   @Id
   private String state;
-  private Double shippingCost;
+  private Float shippingCost;
 
 }
